chatbot_rewe_group/services/streamlit/requirements.txt
streamlit==1.42.1
streamlit-float==0.3.5
streamlit-js-eval==0.1.7
requests==2.32.3
python-dotenv==1.1.1
azure-identity==1.23.0
azure-storage-blob==12.25.1
azure-data-tables==12.7.0
PyYAML==6.0.2
